# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.unknown_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"RecordId": {
				"type": "string",
				"example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
			},
...
			"QuestionAnswerRole": {
				"enum": [
					"questioner",
					"responder"
				],
				"type": "string"
			},
			"QuestionAnswerState": {
				"description": "QuestionAnswer states inferred from RFC 0113.",
				"enum": [
					"question-sent",
					"question-received",
					"answer-received",
					"answer-sent"
				],
				"type": "string"
			},
			"ValidResponse": {
				"properties": {
					"text": {
						"type": "string"
					}
				},
				"required": [
					"text"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_response.string_": {
				"properties": {
					"response": {
						"type": "string"
					}
				},
				"required": [
					"response"
				],
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "credo-controller",
		"version": "0.9.4",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/multi-tenancy/question-answer/{tenantId}": {
			"get": {
				"operationId": "GetQuestionAnswerRecords",
				"responses": {
					"200": {
						"description": "QuestionAnswerRecord[]",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Record_string.unknown_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieve question and answer records by query",
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "Tenant identifier",
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Connection identifier",
						"in": "query",
						"name": "connectionId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Role of the question",
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/QuestionAnswerRole"
						}
					},
					{
						"description": "State of the question",
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/QuestionAnswerState"
						}
					},
					{
						"description": "Thread identifier",
						"in": "query",
						"name": "threadId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/multi-tenancy/question-answer/question/{connectionId}/{tenantId}": {
			"post": {
				"operationId": "SendQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Send a question to a connection",
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "Connection identifier",
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"description": "Tenant identifier",
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"detail": {
										"type": "string"
									},
									"validResponses": {
										"items": {
											"$ref": "#/components/schemas/ValidResponse"
										},
										"type": "array"
									},
									"question": {
										"type": "string"
									}
								},
								"required": [
									"validResponses",
									"question"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/multi-tenancy/question-answer/answer/{id}/{tenantId}": {
			"post": {
				"operationId": "SendAnswer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Send a answer to question",
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "Question Answer Record identifier",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"description": "Tenant identifier",
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Record_response.string_"
							}
						}
					}
				}
			}
		},
		"/multi-tenancy/question-answer/{id}/{tenantId}": {
			"get": {
				"operationId": "GetQuestionAnswerRecordById",
				"responses": {
					"200": {
						"description": "ConnectionRecord",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Retrieve question answer record by id",
				"tags": [
					"MultiTenancy"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "Question Answer Record identifier",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"description": "Tenant identifier",
						"in": "path",
						"name": "tenantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/question-answer": {
			"get": {
				"operationId": "GetQuestionAnswerRecords",
				"responses": {
					"200": {
						"description": "QuestionAnswerRecord[]",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Record_string.unknown_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieve question and answer records by query",
				"tags": [
					"Question Answer"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "Connection identifier",
						"in": "query",
						"name": "connectionId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Role of the question",
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/QuestionAnswerRole"
						}
					},
					{
						"description": "State of the question",
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/QuestionAnswerState"
						}
					},
					{
						"description": "Thread identifier",
						"in": "query",
						"name": "threadId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/question-answer/question/{connectionId}": {
			"post": {
				"operationId": "SendQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.unknown_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"description": "Send a question to a connection",
				"tags": [
					"Question Answer"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "Connection identifier",
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"detail": {
										"type": "string"
									},
									"validResponses": {
										"items": {
											"$ref": "#/components/schemas/ValidResponse"
										},
										"type": "array"
									},
									"question": {
										"type": "string"
									}
								},
								"required": [
									"validResponses",
									"question"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/question-answer/answer/{id}": {
			"post": {
				"operationId": "SendAnswer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.unknown_"
								}
							}
						}
					}
				},
				"description": "Send a answer to question",
				"tags": [
					"Question Answer"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "The id of the question answer record",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Record_response.string_"
							}
						}
					}
				}
			}
		},
		"/question-answer/{id}": {
			"get": {
				"operationId": "GetQuestionAnswerRecordById",
				"responses": {
					"200": {
						"description": "ConnectionRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.unknown_"
								}
							}
						}
					}
				},
				"description": "Retrieve question answer record by id",
				"tags": [
					"Question Answer"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}